fixed bigendianreader double parse bug. fixed issue starting coroutine from non main thread in ValuePublisher

+++ src/Assets/FirstHolo/Scripts/RobotConnect/BigEndianReader.cs
@@ -54,7 +54,7 @@ class NetworkEndianReader : BinaryReader
     {
         a64 = base.ReadBytes(8);
         if (BitConverter.IsLittleEndian) Array.Reverse(a64);
-        return BitConverter.ToSingle(a64, 0);
+        return BitConverter.ToDouble(a64, 0);
     }
 
 }
